Enforce user timezone when converting Unix time
Pocket stores datetime values in Unix (e.g, time when item was added), which the R script converts to a Roam date. By default, `as.POSIXct()` converts into the machine's timezone - in the case of GitHub workflows, they're run in GMT/UTC. For datetimes which are close to midnight (on either side), this can offset the converted date by a day -- so now the timezone is explicitly set in `functions.R` to avoid unexpected behaviors.